backend/requirements.txt
fastapi
uvicorn[standard]
pymupdf==1.25.5
sentence-transformers==4.0.2
transformers==4.48.0
huggingface-hub==0.26.5
faiss-cpu==1.11.0
pillow==11.2.1
pydantic==2.11.3
sqlalchemy==2.0.40
numpy==2.0.2
python-multipart
aiofiles
asyncpg==0.30.0
nltk>=3.9.1
rank_bm25>=0.2.2
sentencepiece
bitsandbytes==0.45.5
accelerate==1.6.0
protobuf==3.20.3